[GUI] Move GUIDimensions to a seperate file
