+++ src/client/gui/gui_rectangle.cpp
@@ -2,20 +2,6 @@
 
 #include <glm/gtc/matrix_transform.hpp>
 
-GuiDimension::GuiDimension(float xScale, float xOffset, float yScale, float yOffset)
-    : scale{xScale, yScale}
-    , offset{xOffset, yOffset}
-{
-}
-
-glm::vec2 GuiDimension::apply(float width, float height) const
-{
-    glm::vec2 vector;
-    vector.x = scale.x * width * 100.0f;
-    vector.y = scale.y * height * 100.0f;
-    return vector + offset;
-}
-
 GuiRectangle::GuiRectangle(const GuiDimension& position, const GuiDimension& size)
     : m_position(position)
     , m_size(size)
