+++ src/client/gui/gui_rectangle.h
@@ -1,17 +1,8 @@
 #pragma once
 
+#include <glm/vec3.hpp>
 #include <glm/glm.hpp>
-#include <glm/vec2.hpp>
-
-struct GuiDimension final {
-    glm::vec2 scale{0.0f};
-    glm::vec2 offset{0.0f};
-
-    GuiDimension() = default;
-    GuiDimension(float xScale, float xOffset, float yScale, float yOffset);
-
-    glm::vec2 apply(float width, float height) const;
-};
+#include "gui_dimension.h"
 
 struct GuiRectangle final {
   public:
